recognize/pyproject.toml
[project]
name = "recognize"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.117.1",
  "ultralytics>=8.3.203",
  "uvicorn>=0.37.0",
  "pillow>=10.0.0",
  "pytest>=8.4.2",
  "torchaudio>=2.8.0",
  "python-multipart>=0.0.20",
  "faststream[rabbit]>=0.5.48",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
